fix : step3 Price 클래스에서 생성 시 validation을 하도록 구현
