+++ src/test/java/step3/LottosGeneratorTest.java
@@ -1,12 +1,10 @@
 package step3;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.ValueSource;
 import step3.domain.Lottos;
 import step3.domain.LottosGenerator;
 import step3.domain.Price;
@@ -19,21 +17,5 @@ public class LottosGeneratorTest {
     void generatePerPrice(final int inputPrice, final int size) {
         final Lottos lottos = LottosGenerator.generateLottos(new Price(inputPrice));
         assertThat(lottos.getLottosCount()).isEqualTo(size);
-    }
-
-    @DisplayName("입력한 금액이 0 혹은 음수면 IllegalArgumentException을 발생시킨다.")
-    @ParameterizedTest
-    @ValueSource(ints = {0, -15000, -12000, -8000})
-    void priceIsMinusThenIllegalArgumentException(final int inputPrice) {
-        assertThatThrownBy(() -> LottosGenerator.generateLottos(new Price(inputPrice)))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("입력한 금액이 천단위로 나누어 떨어지지 않으면 IllegalArgumentException을 발생시킨다.")
-    @ParameterizedTest
-    @ValueSource(ints = {18200, 13250, 1100, 3200})
-    void priceIsNotDivisibleThenIllegalArgumentException(final int inputPrice) {
-        assertThatThrownBy(() -> LottosGenerator.generateLottos(new Price(inputPrice)))
-                .isInstanceOf(IllegalArgumentException.class);
     }
 }
